chore: update version to 0.1.168 and remove unused analysis module
- Bump version in Cargo.toml from 0.1.167 to 0.1.168.
- Remove the `analysis.rs` module as it is no longer needed, streamlining the codebase.
- Update references in `lib.rs` to ensure compatibility with the removal of the analysis module.

These changes enhance maintainability by eliminating obsolete code and updating the version for clarity.
