+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "neat_ai_discovery"
-version = "0.1.167"
+version = "0.1.168"
 edition = "2021"
 
 [lib]
